requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
feedparser==6.0.2
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
sgmllib3k==1.0.0
sqlparse==0.4.1
whitenoise==5.2.0
django-heroku==0.3.1
dj-database-url==0.5.0